Added Type fo toCSV Function in react-csv
