+++ types/react-csv/components/CommonPropTypes.d.ts
@@ -1,12 +1,6 @@
 import { MouseEventHandler } from "react";
+import { Data, Headers } from "../lib/core";
 
-export interface LabelKeyObject {
-    label: string;
-    key: string;
-}
-
-export type Data = object[];
-export type Headers = LabelKeyObject[] | string[];
 export type SyncClickHandler = (event: MouseEventHandler<HTMLAnchorElement>) => boolean | void;
 export type AsyncClickHandler = (event: MouseEventHandler<HTMLAnchorElement>, done: (proceed?: boolean) => void) => void;
