+++ types/react-csv/lib/core.d.ts
@@ -1,0 +1,8 @@
+export interface LabelKeyObject {
+    label: string;
+    key: string;
+}
+
+export type Data = object[];
+export type Headers = LabelKeyObject[] | string[];
+export type toCsv = (data: object[], headers: Headers, separator?: string | undefined, enclosingCharacter?: string | undefined) => void;
