+++ types/react-csv/react-csv-tests.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { render } from "react-dom";
 import { CSVLink, CSVDownload } from "react-csv";
+import { toCSV } from "react-csv/lib/core";
 
 const headers = [
     { label: "First Name", key: "details.firstName" },
@@ -259,3 +260,5 @@ render(
     />,
     node
 );
+
+toCSV(data,headers,",",`"`)
